requirements.txt
# Runtime dependencies
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.70.0
tabulate>=0.9.0

# Dev / testing
pytest>=7.0.0

fastapi==0.104.1
uvicorn[standard]==0.24.0